MRD-2687 Move date formatting file
