+++ integration_tests/integration/contactHistory.spec.ts
@@ -2,7 +2,7 @@ import { DateTime } from 'luxon'
 import getCaseContactHistoryResponse from '../../api/responses/get-case-contact-history.json'
 import { europeLondon, sortListByDateField } from '../../server/utils/dates'
 import { routeUrls } from '../../server/routes/routeUrls'
-import { formatDateTimeFromIsoString } from '../../server/utils/dates/format'
+import { formatDateTimeFromIsoString } from '../../server/utils/dates/format/format'
 import { dedupeList } from '../../server/utils/lists'
 import { removeSystemGenerated } from '../../server/controllers/caseSummary/contactHistory/filterContactsBySystemGenerated'
 import completeRecommendationResponse from '../../api/responses/get-recommendation.json'
