+++ server/controllers/caseSummary/contactHistory/filterContactsByDateRange.ts
@@ -5,7 +5,7 @@ import { convertGmtDatePartsToUtc } from '../../../utils/dates/convert'
 import { ContactSummaryResponse } from '../../../@types/make-recall-decision-api'
 import { dateHasError, europeLondon } from '../../../utils/dates'
 import { formatValidationErrorMessage, invalidDateInputPart, makeErrorObject } from '../../../utils/errors'
-import { formatDateRange } from '../../../utils/dates/format'
+import { formatDateRange } from '../../../utils/dates/format/format'
 import { removeParamsFromQueryString } from '../../../utils/utils'
 import { ContactHistoryFilters } from '../../../@types/contacts'
 import { NamedFormError } from '../../../@types/pagesForms'
