+++ server/controllers/recommendation/ppcs/checkBookingDetailsController.test.ts
@@ -3,14 +3,14 @@ import { mockNext, mockReq, mockRes } from '../../../middleware/testutils/mockRe
 import { getRecommendation, searchForPrisonOffender, updateRecommendation } from '../../../data/makeDecisionApiClient'
 import checkBookingDetailsController from './checkBookingDetailsController'
 import recommendationApiResponse from '../../../../api/responses/get-recommendation.json'
-import { formatDateTimeFromIsoString } from '../../../utils/dates/format'
+import { formatDateTimeFromIsoString } from '../../../utils/dates/format/format'
 import { determinePpudEstablishment } from './determinePpudEstablishment'
 import { randomEnum } from '../../../@types/enum.testFactory'
 import { CUSTODY_GROUP } from '../../../@types/make-recall-decision-api/models/ppud/CustodyGroup'
 import { getRoute } from './custodyGroupRouter'
 
 jest.mock('../../../data/makeDecisionApiClient')
-jest.mock('../../../utils/dates/format')
+jest.mock('../../../utils/dates/format/format')
 jest.mock('./determinePpudEstablishment')
 jest.mock('./custodyGroupRouter')
