+++ server/middleware/setUpMaintenance.ts
@@ -3,7 +3,7 @@ import express, { Router } from 'express'
 import config from '../config'
 import { isDateTimeRangeCurrent } from '../utils/utils'
 import logger from '../../logger'
-import { formatDateTimeFromIsoString } from '../utils/dates/format'
+import { formatDateTimeFromIsoString } from '../utils/dates/format/format'
 
 export default function setUpMaintenance(): Router {
   const router = express.Router()
